requirements.txt
Flask>=3.1,<4.0
APScheduler>=3.10,<4.0
google-api-python-client>=2.170.0,<3.0
google-auth>=2.40.0,<3.0
google-auth-oauthlib>=1.2.0,<2.0
google-auth-httplib2>=0.2.0,<1.0
google-genai>=1.29.0,<2.0
python-dotenv>=1.1.0,<2.0
